+++ internal/model/goal.go
@@ -10,7 +10,7 @@ type Goal struct {
 	Name       string        `json:"name"`
 	AccountIds pq.Int64Array `json:"account_ids" gorm:"type:int[]"`
 	Leader     uint          `json:"leader"`
-	Money      float32       `json:"money"`
+	Money      float64       `json:"money"`
 	Type       int           `json:"type"`
 }
 
@@ -20,7 +20,7 @@ func (m *Goal) TableName() string {
 
 type GoalSetDTO struct {
 	Id         uint    `json:"id"  binding:"required"`
-	Money      float32 `json:"money" binding:"required"`
+	Money      float64 `json:"money" binding:"required"`
 	AccountIds []int64 `json:"account_ids" binding:"required"`
 	Type       int     `json:"type" binding:"required"`
 	Name       string  `json:"name"`
@@ -28,15 +28,15 @@ type GoalSetDTO struct {
 
 type GoalCreateDTO struct {
 	Name  string  `json:"Name" binding:"required"`
-	Money float32 `json:"money" binding:"required"`
+	Money float64 `json:"money" binding:"required"`
 	Type  int     `json:"type" binding:"required"`
 }
 
 type GoalGetDTO struct {
 	Id         uint
 	Name       string
-	Goal       float32
-	CurrMoney  float32
+	Goal       float64
+	CurrMoney  float64
 	Type       int
 	AccountIds []int64
 }
